Move base.enums to base.models.enums
