+++ base/models/exam_enrollment.py
@@ -34,8 +34,7 @@ from django.core.validators import MaxValueValidator, MinValueValidator
 from base.models import person, person_address, session_exam_calendar, session_exam_deadline, \
                         academic_year as academic_yr, offer_year, program_manager, tutor
 from attribution.models import attribution
-from base.enums import exam_enrollment_justification_type as justification_types
-from base.enums import exam_enrollment_state as enrollment_states
+from base.models.enums import exam_enrollment_state as enrollment_states, exam_enrollment_justification_type as justification_types
 
 from base.models.exceptions import JustificationValueException
 from base.models.utils.admin_extentions import remove_delete_action
