+++ base/tests/models/test_person.py
@@ -29,7 +29,7 @@ import factory
 from django.test import TestCase
 from django.test import override_settings
 from base.models import person
-from base.enums import person_source_type
+from base.models.enums import person_source_type
 from base.tests.factories.person import PersonFactory, generate_person_email
